--- a/servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/CompositeException.java
+++ b/servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/CompositeException.java
@@ -22,7 +22,7 @@ import static io.servicetalk.utils.internal.PlatformDependent.throwException;
 
 /**
  * A {@link RuntimeException} that allows to add {@link Throwable} instances at a lower cost than
- * {@link #addSuppressed(Throwable)}. {@link #addAllPendingSuppressed()} will add all pending {@link Throwable}s as
+ * {@link #addSuppressed(Throwable)}. {@link #finishAndThrow()} will add all pending {@link Throwable}s as
  * {@link #addSuppressed(Throwable)}.
  */
 final class CompositeException extends RuntimeException {
@@ -40,7 +40,7 @@ final class CompositeException extends RuntimeException {
 
     /**
      * Add a {@link Throwable} to be added as {@link #addSuppressed(Throwable)} on the next call to
-     * {@link #addAllPendingSuppressed()}.
+     * {@link #finishAndThrow()}.
      *
      * @param toAdd {@link Throwable} to finally add as {@link #addSuppressed(Throwable)}.
      */
@@ -58,7 +58,7 @@ final class CompositeException extends RuntimeException {
      * <p>
      * It is assumed that {@link #add(Throwable)} won't be called after this method.
      */
-    void addAllPendingSuppressed() {
+    void finishAndThrow() {
         Throwable delayedCause = null;
         Throwable next;
         while ((next = suppressed.poll()) != null) {