@Scottmitch 2020-04-18
The benchmarks are not all-encompassing, but provide some indication with reasonable sizes that we don't regress in these scenarios.

```
# This PR (ConcurrentLinkedQueue)
Benchmark                                      Mode  Cnt        Score        Error  Units
CompletableProcessorBenchmark.mtAdd20Complete  thrpt    5   599388.080 ±  5997.789  ops/s
CompletableProcessorBenchmark.mtAdd3Cancel    thrpt    5   111654.849 ±   2162.139  ops/s
CompletableProcessorBenchmark.mtAdd3Complete  thrpt    5   266685.059 ±  16799.889  ops/s
CompletableProcessorBenchmark.stAdd20Complete  thrpt    5  1271502.888 ± 27506.922  ops/s
CompletableProcessorBenchmark.stAdd3Cancel    thrpt    5  6059397.450 ± 145487.879  ops/s
CompletableProcessorBenchmark.stAdd3Complete  thrpt    5  7197153.816 ± 149687.010  ops/s

# master (MPSC + atomics)
Benchmark                                      Mode  Cnt        Score        Error  Units
CompletableProcessorBenchmark.mtAdd20Complete  thrpt    5   270358.067 ± 14923.867  ops/s
CompletableProcessorBenchmark.mtAdd3Cancel    thrpt    5   110551.043 ±   8426.320  ops/s
CompletableProcessorBenchmark.mtAdd3Complete  thrpt    5   194364.989 ±   9631.398  ops/s
CompletableProcessorBenchmark.stAdd20Complete  thrpt    5  1473056.347 ± 38542.670  ops/s
CompletableProcessorBenchmark.stAdd3Cancel    thrpt    5  5131359.228 ± 206785.670  ops/s
CompletableProcessorBenchmark.stAdd3Complete  thrpt    5  6474388.537 ± 211088.908  ops/s
```